model_template/examples/qwen_vl/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pyyaml>=6.0.0
pillow>=10.0.0
httpx>=0.27.0
